+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.14.8] - (01.06.2024)
+
+## Changes
+
+- fix 401 error after envoy reboot and refresh cocies
+- fix characteristics warning for arf profile
+- increase time for check token expired
+= bump dependencies
+- cleanup
+
 ## [7.14.5] - (12.05.2024)
 
 ## Changes